requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-google-genai>=2.0,<3.0
langchain-chroma>=0.1,<0.2
chromadb>=0.4.18
google-generativeai>=0.8.0,<0.9.0  # Updated version range
pandas==2.1.3
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.0.1
pytesseract==0.3.10
Pillow==10.1.0
mysql-connector-python==8.2.0
websockets==12.0

## pip install googleapis-common-protos google-api-core google-ai-generativelanguage grpcio-status

